+++ common/src/main/java/de/maxhenkel/voicechat/gui/widgets/ListScreenBase.java
@@ -1,0 +1,28 @@
+package de.maxhenkel.voicechat.gui.widgets;
+
+import com.mojang.blaze3d.vertex.PoseStack;
+import de.maxhenkel.voicechat.gui.VoiceChatScreenBase;
+import net.minecraft.network.chat.Component;
+
+public class ListScreenBase extends VoiceChatScreenBase {
+
+    private Runnable postRender;
+
+    public ListScreenBase(Component title, int xSize, int ySize) {
+        super(title, xSize, ySize);
+    }
+
+    @Override
+    public void render(PoseStack poseStack, int mouseX, int mouseY, float delta) {
+        super.render(poseStack, mouseX, mouseY, delta);
+        if (postRender != null) {
+            postRender.run();
+            postRender = null;
+        }
+    }
+
+    public void postRender(Runnable postRender) {
+        this.postRender = postRender;
+    }
+
+}