+++ forge/changelog.md
@@ -1,3 +1,5 @@
 - Reworked volume adjustment GUI
+- Reworked join group GUI
+- Reworked group GUI
 - Added system sounds slider
 - Improved GUI performance